+++ XYZReader/src/main/java/com/example/xyzreader/data/repo/BookRepository.java
@@ -72,7 +72,7 @@ public class BookRepository {
         });
     }
 
-    public LiveData<Book> getBook(int id) {
+    public LiveData<Book> getBook(long id) {
         return mBookDao.getBook(id);
     }
